+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.3.1] - 2020-10-19
+
+- Add pyx file to sdist
+- Use Github Actions for CI testing
+
+## [0.3.0] - 2020-10-05
+
+- Allow 2D input for both `positions` and `indices`
+
+## [0.2.2] - 2020-07-10
+
+- Don't build wheels for PyPy. See https://github.com/joerick/cibuildwheel/issues/402
+
+## [0.2.1] - 2020-07-09
+
+- Try to rebuild wheels
+
+## [0.2.0] - 2020-07-01
+
+- New methods for creating a bounding sphere. #10
+
 ## [0.1.2] - 2020-06-01
 
 - Try again to publish to PyPI directly